refa: Piece를 상속 받는 추상클래스 UnrestrictedPiece와 RestrictedPiece 생성
