+++ src/main/java/domain/piece/Piece.java
@@ -3,7 +3,6 @@ package domain.piece;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
@@ -23,18 +22,15 @@ public abstract class Piece {
         this.camp = camp;
     }
 
+    protected abstract void validateMovable(List<Integer> gaps);
+
     public boolean isWhite() {
         return camp.equals(Camp.WHITE);
     }
 
+    public abstract List<Square> fetchMovePath(Square currentSquare, Square targetSquare);
 
-    public List<Square> fetchMovePath(Square currentSquare, Square targetSquare) {
-        List<Integer> gaps = calculateGap(currentSquare, targetSquare);
-        Integer distance = calculateDistance(gaps);
-        validateMovable(gaps);
-        List<Integer> direction = calculateDirection(gaps, distance);
-        return calculatePath(currentSquare, distance, direction);
-    }
+    public abstract boolean canMove(Map<Square, Camp> pathInfo, Square targetSquare);
 
     protected Integer calculateDistance(List<Integer> gaps) {
         List<Integer> absGaps = gaps.stream()
@@ -43,8 +39,6 @@ public abstract class Piece {
         return Collections.max(absGaps);
     }
 
-    protected abstract void validateMovable(List<Integer> gaps);
-
     protected List<Integer> calculateGap(Square currentSquare, Square targetSquare) {
         return IntStream.range(0, 2)
                 .map(i -> targetSquare.toCoordinate().get(i) - currentSquare.toCoordinate().get(i))
@@ -52,46 +46,13 @@ public abstract class Piece {
                 .collect(Collectors.toList());
     }
 
-    private static List<Integer> calculateDirection(List<Integer> gaps, Integer distance) {
-        return gaps.stream()
-                .mapToInt(gap -> gap / distance)
-                .boxed()
-                .collect(Collectors.toList());
-    }
-
-    private List<Square> calculatePath(Square currentSquare, Integer distance, List<Integer> direction) {
-        List<Integer> coordinate = currentSquare.toCoordinate();
-        Integer currentFile = coordinate.get(FILE);
-        Integer currentRank = coordinate.get(RANK);
-        return IntStream.range(1, distance + 1)
-                .mapToObj(dist -> new Square(currentFile + direction.get(FILE) * dist,
-                        currentRank + direction.get(RANK) * dist))
-                .collect(Collectors.toUnmodifiableList());
-    }
-
-    protected boolean isNotForward(List<Integer> gap) {
-        Integer fileGap = gap.get(FILE);
-        Integer RankGap = gap.get(RANK);
-        return fileGap != 0 && RankGap != 0;
-    }
-
-    protected boolean isNotDiagonal(List<Integer> gap) {
-        return !Objects.equals(Math.abs(gap.get(FILE)), Math.abs(gap.get(RANK)));
-    }
-
-    public boolean canMove(Map<Square, Camp> pathInfo, Square targetSquare) {
-        Camp targetCamp = pathInfo.get(targetSquare);
-        pathInfo.remove(targetSquare);
-        return isDifferentCampOrEmptyOnTarget(targetCamp) && !isExistPieceOnPath(pathInfo);
-    }
-
-    private boolean isDifferentCampOrEmptyOnTarget(Camp targetCamp) {
-        return isDifferentCamp(targetCamp) || targetCamp.equals(Camp.NONE);
-    }
-
     protected boolean isExistPieceOnPath(Map<Square, Camp> pathInfo) {
         return pathInfo.values().stream()
                 .anyMatch(camp -> camp != Camp.NONE);
+    }
+
+    protected boolean isDifferentCampOrEmptyOnTarget(Camp targetCamp) {
+        return isDifferentCamp(targetCamp) || targetCamp.equals(Camp.NONE);
     }
 
     protected boolean isDifferentCamp(Camp otherCamp) {
@@ -126,10 +87,6 @@ public abstract class Piece {
     }
 
     public boolean isKing() {
-        return false;
-    }
-
-    public boolean isEmpty() {
         return false;
     }
 }
