+++ src/test/java/domain/piece/type/BishopTest.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.api.Test;
 
 import domain.board.Square;
 import domain.piece.Camp;
+import domain.piece.type.unrestricted.Bishop;
 
 class BishopTest {
