+++ src/test/java/domain/piece/type/QueenTest.java
@@ -2,13 +2,13 @@ package domain.piece.type;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import domain.board.Square;
 import domain.piece.Camp;
+import domain.piece.type.unrestricted.Queen;
 
 class QueenTest {
